requirements.txt
# Core
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# Blog Engine
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
python-wordpress-xmlrpc>=2.3
markdown>=3.5.0

# Video Engine
edge-tts>=6.1.0
moviepy>=1.0.3
Pillow>=10.0.0

# YouTube Upload
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# Micro-SaaS
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# Scheduling & Monitoring
apscheduler>=3.10.0
python-telegram-bot>=20.6

# Database (lightweight)
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
loguru>=0.7.0
tenacity>=8.2.0
